+++ src/Bus/WithCorrelationId.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Butler\Audit\Bus;
+
+use Butler\Audit\Jobs\Middleware\SetCorrelationId;
+
+trait WithCorrelationId
+{
+    public $correlationId;
+
+    public function middleware()
+    {
+        return [new SetCorrelationId()];
+    }
+}