+++ tests/ServiceProviderTest.php
@@ -2,12 +2,23 @@
 
 namespace Butler\Audit\Tests;
 
+use Butler\Audit\Bus\Dispatcher;
 use Butler\Audit\Facades\Auditor;
+use Butler\Audit\Tests\JobWithoutCorrelationId;
 use GrahamCampbell\TestBenchCore\ServiceProviderTrait;
+use Illuminate\Bus\Dispatcher as BaseDispatcher;
+use Illuminate\Queue\Events\JobProcessed;
 
 class ServiceProviderTest extends AbstractTestCase
 {
     use ServiceProviderTrait;
+
+    protected function setUp(): void
+    {
+        putenv('APP_RUNNING_IN_CONSOLE=true');
+
+        parent::setUp();
+    }
 
     public function test_auditor_is_injectable()
     {
@@ -28,5 +39,21 @@ class ServiceProviderTest extends AbstractTestCase
             'secret',
             $this->app->config->get('butler.audit.token')
         );
+    }
+
+    public function test_Dispatcher_is_extended()
+    {
+        $this->assertInstanceOf(Dispatcher::class, app(BaseDispatcher::class));
+    }
+
+    public function test_correlation_id_is_reset_after_each_queued_job()
+    {
+        $this->assertTrue(app('events')->hasListeners(JobProcessed::class));
+
+        $correlationId = Auditor::correlationId();
+
+        event(new JobProcessed('connection', new JobWithoutCorrelationId()));
+
+        $this->assertNotEquals($correlationId, Auditor::correlationId());
     }
 }