+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- **BREAKING:**: Handle correlation id for queued job by extending job "Dispatcher" and adding "WithCorrelationId" trait.
+
 ### Changed
 - **BREAKING:**: Require PHP 8.0.